LayoutTests/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-editing-host-cannot-have-div.tentative_host=p&white-space=normal&display=block&command=insertText-expected.txt

a
b


FAIL <p contenteditable style="display:block; white-space:normal">a[]b</p> assert_equals: expected "a<br>b" but got "a<div>b</div>"
FAIL <p contenteditable style="display:block; white-space:normal"><span style="white-space:normal">a[]b</span></p> assert_equals: expected "<span style=\"white-space:normal\">a<br>b</span>" but got "<span style=\"white-space:normal\">a</span><div><span style=\"white-space:normal\">b</span></div>"
PASS <p contenteditable style="display:block; white-space:normal"><span style="display:block;white-space:normal">a[]b</span></p>
PASS <p contenteditable style="display:block; white-space:normal"><div>a[]b</div></p>
FAIL <p contenteditable style="display:block; white-space:normal"><div style="display:inline">a[]b</div></p> assert_equals: expected "<div style=\"display:inline\">a\nb</div>" but got "<div style=\"display:inline\">a</div><div><div style=\"display:inline\">b</div></div>"
FAIL <p contenteditable style="display:block; white-space:normal"><div style="display:inline-block">a[]b</div></p> assert_equals: expected "<div style=\"display:inline-block\">a<br>b</div>" but got "<div style=\"display:inline-block\">a</div><div><div style=\"display:inline-block\">b</div></div>"
FAIL <p contenteditable style="display:block; white-space:normal"><div style="display:inline;white-space:normal">a[]b</div></p> assert_equals: expected "<div style=\"display:inline;white-space:normal\">a<br>b</div>" but got "<div style=\"display:inline;white-space:normal\">a</div><div><div style=\"display:inline;white-space:normal\">b</div></div>"
FAIL <p contenteditable style="display:block; white-space:normal"><div style="display:inline-block;white-space:normal">a[]b</div></p> assert_equals: expected "<div style=\"display:inline-block;white-space:normal\">a<br>b</div>" but got "<div style=\"display:inline-block;white-space:normal\">a</div><div><div style=\"display:inline-block;white-space:normal\">b</div></div>"
